Checked much more strict compile flags and removed nearly all compiler warnings. (on OS X)
git-svn-id: http://anonsvn.internet2.edu/svn/I2util/branches/BWCTL_1_3rc1@202 af504c07-53ac-4b46-a7f1-a0422916b9a2

+++ I2util/io.h
@@ -61,7 +61,7 @@ I2CopyFile(
         I2ErrHandle eh,
         int         tofd,
         int         fromfd,
-        off_t       len
+        size_t      len
         );
 END_C_DECLS
 #endif	/*	_i2_io_h_	*/
